backend/pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Artyom", email = "artyom.groshev2017@gmail.com"},
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "fastapi>=0.89.1",
    "uvicorn>=0.20.0",
    "pydantic>=1.10.4",
    "email-validator>=1.3.1",
    "psycopg[binary]>=3.1.8",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "redis>=4.4.2",
    "python-dateutil>=2.8.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
